Revert "BackService completed"
This reverts commit 5c928ed7263d32e58891a8e9a10e21e5f34b5158.
