--- Independent.Services/Template10.Services.BackService/Nuget/Template10.Services.BackService.nuspec
@@ -1,43 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>Template10.Services.BackService</id>
-    <version>1.0.0</version>
-    <title>Template10.Services.BackService</title>
-    <authors>Jerry Nixon</authors>
-    <owners>Jerry Nixon</owners>
-    <licenseUrl>http://aka.ms/Template10-License</licenseUrl>
-    <projectUrl>http://aka.ms/Template10-Nuget</projectUrl>
-    <iconUrl>https://raw.githubusercontent.com/Windows-XAML/Template10/master/Assets/NuGetLogo.png</iconUrl>
-
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-
-    <description>
-      <![CDATA[ Abstract service that seemlessly handles back navigation for Template10]]>
-    </description>
-
-    <copyright>Jerry Nixon (c) 2015-2017</copyright>
-     
-    <language />
-
-    <tags> Windows 10, Windows10, W10, UAP, UWP, WinRT, XAML, Template 10, Template10, T10, Windows Universal, Universal App</tags>
-
-
-    <dependencies  >
-      <group targetFramework="uap">
-      	<dependency id="Microsoft.NETCore.UniversalWindowsPlatform" version="5.3.3" />
-      </group>
-    </dependencies>
-    
-    <releaseNotes>
-        Version 1.0.0
-        First Release
-    </releaseNotes>
-  </metadata>
-
-  <files>
-    <!--<file src="lib\" target="lib\uap10.0\" />-->
-    <file src="bin\$configuration$\*.dll" target="lib\uap10.0\" />
-    <file src="bin\$configuration$\*.pri" target="lib\uap10.0\" />
-  </files>
-</package>
