+++ Template10/Template10.NavigationService/project.json
@@ -1,8 +1,7 @@
 ﻿{
   "dependencies": {
     "Microsoft.NETCore.UniversalWindowsPlatform": "5.3.3",
-    "Microsoft.Xaml.Behaviors.Uwp.Managed": "2.0.0",
-    "Template10.Services.KeyboardService": "1.0.0"
+    "Microsoft.Xaml.Behaviors.Uwp.Managed": "2.0.0"
   },
   "frameworks": {
     "uap10.0": {}
